+++ server/src/jobs/seed/fixtures/defaultRolesAcls.js
@@ -62,7 +62,8 @@ module.exports = {
       "dossier/voir_contrat_pdf",
       "dossier/voir_contrat_pdf/telecharger",
       "dossier/page_signatures",
-      "dossier/page_signatures/signer",
+      "dossier/page_signatures/signature_papier",
+      "dossier/page_signatures/signature_electronique",
       "dossier/page_statut",
       "dossier/publication",
       "dossier/envoyer",
@@ -86,7 +87,8 @@ module.exports = {
       "dossier/voir_contrat_pdf",
       "dossier/voir_contrat_pdf/telecharger",
       "dossier/page_signatures",
-      "dossier/page_signatures/signer",
+      "dossier/page_signatures/signature_papier",
+      "dossier/page_signatures/signature_electronique",
       "dossier/page_statut",
       "dossier/publication",
       "dossier/envoyer",
@@ -126,11 +128,4 @@ module.exports = {
     description: "Rôle utilisateur Organisme de formation",
     acl: [],
   },
-  // instructeur: {
-  //   name: "instructeur",
-  //   type: "user",
-  //   title: "Instructeur",
-  //   description: "Rôle utilisateur DR(I)EETS",
-  //   acl: [],
-  // },
 };
