+++ ui/components/Workspace/TableDossiers.jsx
@@ -47,7 +47,14 @@ const TableDossiers = ({
       }))}
       components={{
         Nom: (value, i) => {
-          let step = "signatures";
+          let step = "suivi";
+          if (
+            dossiers[i].etat === "EN_ATTENTE_SIGNATURES" ||
+            dossiers[i].etat === "DOSSIER_FINALISE_EN_ATTENTE_ACTION" ||
+            dossiers[i].etat === "SIGNATURES_EN_COURS" ||
+            dossiers[i].etat === "EN_ATTENTE_DECLENCHEMENT_SIGNATURES"
+          )
+            step = "signatures";
           if (dossiers[i].etat === "BROUILLON") step = "cerfa";
 
           return (
