+++ ui/package.json
@@ -43,6 +43,7 @@
     "react-dom": "17.0.2",
     "react-dropzone": "11.4.2",
     "react-focus-lock": "2.7.1",
+    "react-icons": "4.4.0",
     "react-imask": "6.2.2",
     "react-markdown": "8.0.0",
     "react-notion-x": "4.12.0",
